fix: split off remaining reconstruction benchmkark plugin code
