+++ src/benchmarks/reconstruction/TRACKINGcheck/TRACKINGcheck.cc
@@ -1,0 +1,9 @@
+#include "TRACKINGcheckProcessor.h"
+
+// The following just makes this a JANA plugin
+extern "C" {
+    void InitPlugin(JApplication *app) {
+        InitJANAPlugin(app);
+        app->Add(new TRACKINGcheckProcessor);
+    }
+}
