+++ src/benchmarks/reconstruction/TRACKINGcheck/TRACKINGcheckProcessor.cc
@@ -9,14 +9,6 @@
 #include <Evaluator/DD4hepUnits.h>
 #include <TVector3.h>
 
-// The following just makes this a JANA plugin
-extern "C" {
-    void InitPlugin(JApplication *app) {
-        InitJANAPlugin(app);
-        app->Add(new TRACKINGcheckProcessor);
-    }
-}
-
 //-------------------------------------------
 // InitWithGlobalRootLock
 //-------------------------------------------
