+++ src/benchmarks/reconstruction/femc_studies/femc_studies.cc
@@ -1,0 +1,9 @@
+#include "femc_studiesProcessor.h"
+
+// The following just makes this a JANA plugin
+extern "C" {
+  void InitPlugin(JApplication* app) {
+    InitJANAPlugin(app);
+    app->Add(new femc_studiesProcessor());
+  }
+}
