+++ src/benchmarks/reconstruction/lfhcal_studies/lfhcal_studies.cc
@@ -1,0 +1,9 @@
+#include "lfhcal_studiesProcessor.h"
+
+// The following just makes this a JANA plugin
+extern "C" {
+  void InitPlugin(JApplication* app) {
+    InitJANAPlugin(app);
+    app->Add(new lfhcal_studiesProcessor());
+  }
+}
