+++ src/benchmarks/reconstruction/lfhcal_studies/lfhcal_studiesProcessor.cc
@@ -34,14 +34,6 @@
 
 #include "clusterizer_MA.h"
 
-// The following just makes this a JANA plugin
-extern "C" {
-  void InitPlugin(JApplication* app) {
-    InitJANAPlugin(app);
-    app->Add(new lfhcal_studiesProcessor());
-  }
-}
-
 
 //******************************************************************************************//
 // InitWithGlobalRootLock
